refactor: QueryClientWrapper 컴포넌트화 후 global wrapping
