+++ apps/client/src/app/address/[type]/page.tsx
@@ -1,13 +1,7 @@
 'use client';
 
-import ReactQueryProvider from '@/providers/QueryProvider';
-
 import AddressTypePageContent from './AddressTypePageContent';
 
 export default function AddressCUPage() {
-    return (
-        <ReactQueryProvider>
-            <AddressTypePageContent />
-        </ReactQueryProvider>
-    );
+    return <AddressTypePageContent />;
 }
