+++ apps/client/src/app/order/form/page.tsx
@@ -1,13 +1,7 @@
 'use client';
 
-import ReactQueryProvider from '@/providers/QueryProvider';
-
 import OrderFormContent from './OrderFormPageContent';
 
 export default function OrderFormPage() {
-    return (
-        <ReactQueryProvider>
-            <OrderFormContent />
-        </ReactQueryProvider>
-    );
+    return <OrderFormContent />;
 }
