+++ apps/client/src/providers/QueryProviderWrapper.tsx
@@ -1,0 +1,9 @@
+import ReactQueryProvider from './QueryProvider';
+
+export default function ReactQueryProviderWrapper({
+    children,
+}: {
+    children: React.ReactNode;
+}) {
+    return <ReactQueryProvider>{children}</ReactQueryProvider>;
+}
